Turn `composer` version into a metric
We'd like to know how many jobs are running using `composer 1` vs
`composer 2`, so this exposes it as a metric.

Long term, this can be further refined to report `composer` versions
with more granularity, for example `2.3`, `2.4`, `2.5`, etc.

But right now we don't have active questions that'd be answered by
knowing that granularity, and our metrics provider charges based on the
number of metrics we emit, so `2` is an appropriate bucket for now.